Move skin component localisations to namespaces
