+++ osu.Game/Localisation/SkinComponents/FontAdjustableSkinComponentStrings.cs
@@ -3,11 +3,11 @@
 
 using osu.Framework.Localisation;
 
-namespace osu.Game.Localisation
+namespace osu.Game.Localisation.SkinComponents
 {
     public static class FontAdjustableSkinComponentStrings
     {
-        private const string prefix = @"osu.Game.Resources.Localisation.FontAdjustableSkinComponent";
+        private const string prefix = @"osu.Game.Resources.Localisation.SkinComponents.FontAdjustableSkinComponent";
 
         /// <summary>
         /// "Font"
